Update to indicate that it is based on US copyright
Please consider updating to indicate that this is based on a US copyright perspective. Rules around the use of copyrighted material (archival or otherwise) and public domain can vary from one country to the next. Overall this is a great, succinct primer when it comes to thinking about reuse of archival material - thank you! And thank you for Tropy - which is also great!